cifsd-tools: rename cifsadmin to cifsuseradd
useradd is a UNIX tool to add or delete users.

cifsadmin is not a tool for administration. It's just a tool
which can add or delete user. And that's it. Rename it to
reflect the tool's purpose, so we'll have a naming scheme
similar to UNIX.

We will have more administration tools later. Each tool
will do one thing only, which is, once again, very UNIX.

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>