cifsd-tools: handle orphaned lock file
When cifsd is terminated with SIGKILL (or in any other ungraceful
way) it cannot handle the signal appropriately and unlink() its
cifsd.lock file. In this case we can end up having an orphaned
cifsd.lock file - file exists, but no cifsd process owns it.

Add a simple code to handle orphaned lock files. We store manager
PID in cifsd.lock file, so we can check if that /proc/%PID% entry
exists. If it does then we, likely, have active cifsd. Otherwise
the file is orphaned and we can just unlink() it.

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>