ksmbd-tools: set STATUS_SOME_NOT_MAPPED response for domain sid stored by other server
If xattr acl meta in file or directory is stored by samba, domain sid
could be different with ksmbd. In this case, the server must send the
STATUS_SOME_NOT_MAPPED response for the client's lsarpc lookup
sid2 request. If a domain sid in request is wrong by comparing own domain
sid in server, Type is set to SID_TYPE_UNKNOWN and return to
STATUS_SOME_NOT_MAPPED.

Signed-off-by: Namjae Jeon <namjae.jeon@samsung.com>