ksmbd-tools: refuse guest sessions if "guest ok = no" for a share
Currently, the "guest ok" behaviour depends on the global "map to user"
setting.

Consider the following config:

[global]
	interfaces = lo eth0
	bind interfaces only = yes
	netbios name = SMBD
	map to guest = never

[ksmbd share]
	path = /tmp
	guest ok = no
	read only = no

With ksmbd.mountd, guest access is denied only if "map to guest = bad
user".
With smbd, guest access is always denied on a share if "guest ok = no".

This patch ensures the "guest ok" setting is consistently respected
regardless of the "map to guest" configuration (and matches samba4
behaviour).

Signed-off-by: Marios Makassikis <mmakassikis@freebox.fr>
Signed-off-by: Namjae Jeon <namjae.jeon@samsung.com>