ksmbd-tools: Use guest account only if no user was specified
If a share enables guest access, tcm_handle_tree_connect() will use the
guest account even if the request contains a username.

With the following config (and an appropriate ksmbd.pwd file):

    [global]
            netbios name = SMBD
            map to guest = never

    [share]
            path = /tmp
            guest ok = yes

One can observe the following behavior:

  $ mount -t cifs //127.0.0.1/share share -o username=example,pass=pass
  $ umount share
  $ mount -t cifs //127.0.0.1/share share -o username=example,pass=pass
  mount error(13): Permission denied
  Refer to the mount.cifs(8) manual page (e.g. man mount.cifs)

The following message is logged by ksmbd.mountd:
  [ksmbd-worker/6444]: DEBUG: Kill user example

This happens because there is an implicit tree connection to the IPC$
share, which belongs to the same session. Since the IPC$ share doesn't
have the KSMBD_SHARE_FLAG_GUEST_OK flag set, different users are refcounted
(guest for /tmp, 'example' for IPC$). However, only one user is saved in
the session structure.

When sm_handle_tree_disconnect() is called, it decrements the refcounter
of the user that was specified when the session was created.

Signed-off-by: Marios Makassikis <mmakassikis@freebox.fr>
Signed-off-by: Namjae Jeon <namjae.jeon@samsung.com>