ksmbd-tools: fix maximum password length when prompting
MAX_NT_PWD_LEN is the maximum length of the password *including* the
null terminator. Despite that, g_try_malloc0() at the beginning of
__prompt_password_stdin() allocates MAX_NT_PWD_LEN+1 bytes. This is
presumably done because of the trailing newline that fgets() inserts
into the buffer. This newline is converted into a new null terminator
later in __prompt_password_stdin(). However, the second argument to
fgets() was MAX_NT_PWD_LEN, which means that at most MAX_NT_PWD_LEN-1
characters are read. fgets() then writes the null terminator to
zero-based index MAX_NT_PWD_LEN-1. This meant that the maximum length
of the password when prompting was actually MAX_NT_PWD_LEN-1
*including* the null terminator. Fix this by changing the second
argument of fgets() to MAX_NT_PWD_LEN+1.

Signed-off-by: atheik <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>