ksmbd-tools: fix maximum password length detection when prompting
When prompting for a password and the maximum password length was
exceeded, an error message was printed about non matching passwords.
Instead, detect when fgets() returns before reading a newline, consume
the extra characters, and reprompt. Also, if an error happens in the
prompting, output a newline to stdout instead of stderr before the
error message, since there is no guarantee that stdout and stderr
both output to the terminal.

Signed-off-by: atheik <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>