ksmbd-tools: fix maximum password length detection with option
Maximum password length was not checked when using the `password'
option of ksmbd.adduser. Add the check and call exit() on error instead
of returning due to the return type not differentiating errors. Also,
notify the user if an empty password is given using the option. Make
this message less ambiguous in both the prompt and the option by
referring to an `empty password' instead of `no password'.

Signed-off-by: atheik <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>