ksmbd-tools: move verbose option to beginning of usage string
Some options that are affected by `--verbose' terminate further
argument handling, namely the `--shutdown` and `--reload` option of
control. This means that for `--verbose' to have an effect, it must be
specified before such options. Move `--verbose' to the beginning of the
usage string in control. For the sake of consistency, do that for the
other utilities as well.

Signed-off-by: atheik <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>