ksmbd-tools: document utf-8 support in user and share names
User names may be UTF-8 instead of ASCII but have the same issue as
password-input, in that they are assumed to be UTF-8 and not converted
from the current locale. The same is true for share names but their
case-insensitivity only works with recent versions of ksmbd. Update the
documentation and usage strings to reflect this. Additionally, make a
few minor changes.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>