ksmbd-tools: lookup both guest accounts on tree connect
When both global guest account and share guest account exist, the ref
count of the former is decremented such that it is killed on guest
session disconnect. This is because a lookup is done only for the
latter when determining whether to give guest access on tree connect.
Fix this by always doing a lookup for both on such tree connects.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>