ksmbd-tools: use 'disabled', 'auto' and 'mandatory' for smb3 encryption parameter
Use 'disabled', 'auto' and 'mandatory' for smb3 encryption parameter
instead of 'yes' or 'no'.

With smb3 encryption = disabled, SMB3 encryption is disallowed even if it
is requested by the client.
With smb3 encryption = auto, SMB3 encryption is allowed if it is requested
by the client.
With smb3 encryption = mandatory, SMB3 encryption is required.

Default : smb3 encryption = auto.

Reviewed-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>