ksmbd-tools: mountd: fix out of bounds access in set_domain_name()
smb_sid_to_string() does not check bounds when writing to its output
buffer. Additionally, it does not null terminate it, so the call to
strlen() in set_domain_name() can go past the end of the buffer.

Replace calls to strcpy() with snprintf() to ensure null termination
and detect overflows.

Signed-off-by: Marios Makassikis <mmakassikis@freebox.fr>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>