ksmbd-tools: fix handling of max connections on config reload
As of commit 785a9d2, `max connections' is both a global parameter and
a share parameter. This introduces a bug where the global parameter
is confused for the share parameter on config reload if a share has not
specified the latter. This bug is identical to the one fixed in commit
a09f767 for the `guest account' global and share parameters. As such,
fix it the same way. Additionally, since the `max connections' *share*
parameter can no longer be given in the global section, remove it from
the ksmbd.conf.example file. Also, fix the `smb3 encryption' default.
Finally, describe both the global parameter and the share parameter in
ksmbd.conf(5) since they are distinct. Omit the use of number suffixes
since they are not supported by the share parameter code path.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>