ksmbd-tools: remove g_print() and g_printerr() usage
Both g_print() and g_printerr() do character set conversion based on
the current locale which is not desirable since UTF-8 is assumed
throughout. There is no known rationale for their use in this codebase.
While we're at it, as there is no established convention, print usage
page to stdout instead of stderr.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>