ksmbd-tools: fix subauth file creation bugs
Stop leaking g_rand_new(). Also, a subauth value is an `unsigned int'
and as such its range should be [0, UINT_MAX] instead of the current
[0, INT_MAX-1]. Remove write_file_safe() and reuse set_conf_contents()
instead as it has similar file consistency guarantees. The permissions
are now 0640. If parsing the file fails, generate random subauths and
use them. If parsing failed due to the file not existing, the generated
subauths will be written, but failing to do so is not an error.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>