ksmbd-tools: introduce load_config() and remove_config()
Since set_conf_contents() always sets the file perms to 0640, there are
few situations where, *when* using the utilities, the file is readable
but not writable. As such, the only significant use case remaining for
test_file_access() is checking whether the file can be created. Remove
test_file_access() and do the check with set_conf_contents(). Also, do
it in the config parser since there is already a precedent for it.

Introduce load_config() and remove_config() that perform the proper
per-tool resource load and final release, respectively.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>